requirements.txt
apache-airflow~=2.10.1
gitpython~=3.1.43